requirements.txt
discord.py
wikipedia
requests
sympy
PyNaCl
mplfinance
matplotlib
pandas